requirements.txt
torch
torchvision
tensorboard
timm
opencv-python
setuptools
easydict
wget
scikit-image
fairscale
einops
ttkbootstrap
pillow